+++ packages/use/index.ts
@@ -3,3 +3,4 @@ export * from './src/whenever/index';
 export * from './src/wheneverEffectScope/index';
 export * from './src/useCountdown/index';
 export * from './src/deepUnref/index';
+export * from './src/useRequest/index';