+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
   - 🌟 [@mixte/use] 新增 `deepUnref` 方法
+  - 🌟 [@mixte/use] 新增 `useRequest` 方法
 
 ## [v1.1.1]
   - 📅 2023-08-31